requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==1.7.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
